+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [this versioning scheme](https://gist.github.com/cech12/69319028e88c50349a6b044000a6607b).
 
+## [1.21.3-1.1.0.0] - 2025-10-26
+### Changed
+- updated to Minecraft 1.21.3 (Fabric 0.114.1+1.21.3, NeoForge 21.3.94, Forge 53.1.0)
+- updated Cloth Config support (16.0.143) (Fabric/Quilt)
+- updated ModMenu support (12.0.0) (Fabric/Quilt)
+- use the new common tag "c:tools/shear" for shears (Forge)
+
 ## [1.21.1-1.0.0.0] - 2025-10-26
 ### Added
 - supported Minecraft 1.21.1 (Fabric 0.116.7+1.21.1, NeoForge 21.1.213, Forge 52.1.0)